sensor-simulation/requirements.txt
paho-mqtt==1.6.1
requests==2.31.0
python-dotenv==1.0.0
numpy==1.26.3
schedule==1.2.1
faker==22.5.1  # For generating realistic sensor names/locations 